requirements.txt
Appium-Python-Client~=4.5.0
pytest~=8.3.4